Correction to Attribute Value for Public Slave
Correcting attribute value for a public slave. Setting node['dcos']['dcos_role'] to 'slave-public' results in an error, in which the installer script suggests "master", "slave" or "slave_public". Setting an attribute of "slave_public" does seem to correct the problem and result in the install and configuration as expected.