Helper: dcos_cluster_address
Include a new helper method to get the address to a DC/OS cluster
for sending API calls or CLI commands. Use `leader.mesos` as the
default address if we're not configured with either an external
load balancer on the masters or a configured master list. This can
be useful for cases where you may want to install local tools to
run against a remote cluster.

Signed-off-by: Chris Gianelloni <wolf31o2@gmail.com>